Refactor alert_manager into its own module Add `Alert.add_context` method Further define AlertManager methods to add/remove alerts
